+++ src/app/providers/StoreProvider/index.ts
@@ -5,6 +5,9 @@ import type { StateSchema, ReduxStoreWithManager, ThunkConfig } from './config/S
 export {
     StoreProvider,
     createReduxStore,
+};
+
+export type {
     StateSchema,
     ReduxStoreWithManager,
     AppDispatch,
