+++ src/entities/Article/index.ts
@@ -1,8 +1,9 @@
 export { ArticleDetails } from './ui/ArticleDetails/ArticleDetails';
-export {
-    Article, ArticleView, ArticleSortField, ArticleType,
+export type {
+    Article,
 } from './model/types/article';
-export { ArticleDetailsSchema } from './model/types/articleDetailsSchema';
+export { ArticleView, ArticleSortField, ArticleType } from './model/consts/consts';
+export type { ArticleDetailsSchema } from './model/types/articleDetailsSchema';
 export { getArticleDetailsData } from './model/selectors/articleDetails';
 export { ArticleList } from './ui/ArticleList/ArticleList';
 export { ArticleViewSelector } from './ui/ArticleViewSelector/ArticleViewSelector';
