+++ src/entities/Article/model/consts/consts.ts
@@ -1,0 +1,23 @@
+export enum ArticleBlockType {
+    CODE = 'CODE',
+    IMAGE = 'IMAGE',
+    TEXT = 'TEXT',
+}
+
+export enum ArticleSortField {
+    VIEWS = 'views',
+    TITLE = 'title',
+    CREATED = 'createdAt',
+}
+
+export enum ArticleType {
+    ALL = 'ALL',
+    IT = 'IT',
+    SCIENCE = 'SCIENCE',
+    ECONOMIC = 'ECONOMIC',
+}
+
+export enum ArticleView {
+    LIST = 'LIST',
+    GRID = 'GRID',
+}
