+++ src/entities/Article/model/services/fetchArticleById/fetchArticleById.test.ts
@@ -1,6 +1,6 @@
 import { TestAsyncThunk } from 'shared/libs/tests/TestAsyncThunk/TestAsyncThunk';
 import { fetcArticleById } from './fetchArticleById';
-import { ArticleBlockType, ArticleType } from '../../types/article';
+import { ArticleBlockType, ArticleType } from '../../consts/consts';
 
 const data = {
     id: '1',
