+++ src/entities/Article/ui/ArticleListItem/ArticleListItemSceleton.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { classNames } from 'shared/libs/classNames/classNames';
 import { Card } from 'shared/ui/Card/Card';
 import { Skeleton } from 'shared/ui/Skeleton/Skeleton';
-import { ArticleView } from '../../model/types/article';
+import { ArticleView } from '../../model/consts/consts';
 import cls from './ArticleListItem.module.scss';
 
 interface ArticleListItemSceletonProps {
