+++ src/entities/Article/ui/ArticleTypeTabs/ArticleTypeTabs.tsx
@@ -2,7 +2,7 @@ import { memo, useCallback, useMemo } from 'react';
 import { classNames } from 'shared/libs/classNames/classNames';
 import { useTranslation } from 'react-i18next';
 import { TabItem, Tabs } from 'shared/ui/Tabs/Tabs';
-import { ArticleType } from '../../model/types/article';
+import { ArticleType } from '../../model/consts/consts';
 
 interface ArticleTypeTabsProps {
     className?: string;
