+++ src/entities/Country/model/consts/county.ts
