+++ src/entities/Country/ui/CountrySelect/CountrySelect.tsx
@@ -2,7 +2,7 @@ import { memo, useCallback } from 'react';
 import { classNames } from 'shared/libs/classNames/classNames';
 import { useTranslation } from 'react-i18next';
 import { ListBox } from 'shared/ui/ListBox/ListBox';
-import { Country } from '../../model/types/county';
+import { Country } from '../../model/consts/county';
 
 interface CountrySelectProps {
     className?: string;
