+++ src/entities/Currency/index.ts
@@ -1,2 +1,2 @@
-export { Currency } from './model/types/currency';
+export { Currency } from './model/consts/currency';
 export { CurrencySelect } from './ui/CurrencySelect/CurrencySelect';
