+++ src/entities/Currency/model/consts/currency.ts
