+++ src/entities/User/index.ts
@@ -1,5 +1,6 @@
 export { userReducer, userActions } from './model/slice/userSlice';
-export { UserSchema, User, UserRole } from './model/types/user';
+export type { UserSchema, User } from './model/types/user';
+export { UserRole } from './model/consts/consts';
 export { getUserAuthData } from './model/selectors/getUserAuthData';
 export { getUserInited } from './model/selectors/getUserInited';
 export { isUserAdmin, isUserManager, getUserRole } from './model/selectors/roleSelectors';
