+++ src/entities/User/model/consts/consts.ts
@@ -1,0 +1,5 @@
+export enum UserRole {
+    ADMIN = 'ADMIN',
+    USER = 'USER',
+    MANAGER = 'MANAGER',
+}
