+++ src/entities/User/model/selectors/roleSelectors.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { StateSchema } from 'app/providers/StoreProvider';
-import { UserRole } from '../types/user';
+import { UserRole } from '../consts/consts';
 
 export const getUserRole = (state: StateSchema) => state.user.authData?.roles;
