+++ src/features/AuthByUsername/index.ts
@@ -1,2 +1,2 @@
 export { LoginModal } from './ui/LoginModal/LoginModal';
-export { LoginSchema } from './model/types/loginSchema';
+export type { LoginSchema } from './model/types/loginSchema';
