+++ src/features/ScrollSave/index.ts
@@ -1,3 +1,3 @@
-export { ScrollSaveSchema } from './model/types/ScrollSaveSchema';
+export type { ScrollSaveSchema } from './model/types/ScrollSaveSchema';
 export { getScrollSaveByPath } from './model/selectors/getScrollSave';
 export { ScrollSaveReducer, ScrollSaveActions } from './model/slices/ScrollSaveSlice';
