+++ src/features/editableProfileCard/index.ts
@@ -1,3 +1,3 @@
 export { EditableProfileCard } from './ui/EditableProfileCard/EditableProfileCard';
-export { ProfileSchema } from './model/types/editableProfileCardSchema';
+export type { ProfileSchema } from './model/types/editableProfileCardSchema';
 export { profileReducer } from './model/slice/profileSlice';
