+++ src/features/editableProfileCard/model/consts/consts.ts
@@ -1,0 +1,8 @@
+export enum ValidateProfileError {
+    INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
+    INCORRECT_AGE = 'INCORRECT_AGE',
+    INCORRECT_CITY = 'INCORRECT_CITY',
+    INCORRECT_USERNAME = 'INCORRECT_USERNAME',
+    NO_DATA = 'NO_DATA',
+    SERVER_ERROR = 'SERVER_ERROR',
+}
