+++ src/features/editableProfileCard/model/selectors/getProfileValidateErrors/getProfileValidateErrors.test.ts
@@ -1,6 +1,6 @@
 import { StateSchema } from 'app/providers/StoreProvider';
 import { getProfileValidateErrors } from './getProfileValidateErrors';
-import { ValidateProfileError } from '../../types/editableProfileCardSchema';
+import { ValidateProfileError } from '../../consts/consts';
 
 describe('getProfileValidateErrors', () => {
     test('should work with filled state', () => {
