+++ src/pages/ArticleDetailsPage/index.ts
@@ -1,5 +1,5 @@
 export { ArticleDetailsPageAsync as ArticleDetailsPage } from './ui/ArticleDetailsPage/ArticleDetailsPage.async';
 
-export { ArticleDetailsCommentsSchema } from './modal/types/ArticleDetailsCommentsSchema';
-export { ArticleDetailsRecommendationSchema } from './modal/types/ArticleDetailsRecommendationSchema';
-export { ArticlesDetailsPageSchema } from './modal/types';
+export type { ArticleDetailsCommentsSchema } from './modal/types/ArticleDetailsCommentsSchema';
+export type { ArticleDetailsRecommendationSchema } from './modal/types/ArticleDetailsRecommendationSchema';
+export type { ArticlesDetailsPageSchema } from './modal/types';
