+++ src/pages/ArticlesPage/index.ts
@@ -1,3 +1,3 @@
 export { ArticlesPageAsync as ArticlesPage } from './ui/ArticlesPage/ArticlesPage.async';
 
-export { ArticlesPageSchema } from './model/types/articlePageSchema';
+export type { ArticlesPageSchema } from './model/types/articlePageSchema';
