+++ tsconfig.json
@@ -25,6 +25,7 @@
         // Если какая-либо библиотека не имеет default import, лоадеры вроде ts-loader или babel-loader автоматически создает их вместо такого импорта import * as React from 'react'; можно писать такой import React from 'react';
         "allowSyntheticDefaultImports": true,
         "noImplicitThis": true,
+        "isolatedModules": true,
     },
     "ts-node": {
         "compilerOptions": {
